Improve compatibility with chef 13. Eliminate cloned resource warnings
