+++ recipes/default.rb
@@ -38,12 +38,12 @@ ruby_block "run-iptables-resources-early" do
     node.rm_normal("simple_iptables", "ipv6")
     # Then run all the simple_iptables_* resources
     run_context.resource_collection.each do |resource|
-      if resource.kind_of?(Chef::Resource::SimpleIptablesRule)
+      if resource.resource_name == :simple_iptables_rule
         Chef::Log.debug("about to run simple_iptables_rule[#{resource.chain}]")
-        resource.run_action(resource.action)
-      elsif resource.kind_of?(Chef::Resource::SimpleIptablesPolicy)
+        resource.run_action(resource.action.first)
+      elsif resource.resource_name == :simple_iptables_policy
         Chef::Log.debug("about to run simple_iptables_policy[#{resource.chain}]")
-        resource.run_action(resource.action)
+        resource.run_action(resource.action.first)
       end
     end
